Move `types` pachage to `internal` subdirectory
