+++ src/kubeobjects/kubernetes/statefulset/modifiers/container_adder.go
@@ -1,7 +1,7 @@
 package modifiers
 
 import (
-	"github.com/Dynatrace/dynatrace-operator/src/kubeobjects/kubernetes/statefulset/types"
+	"github.com/Dynatrace/dynatrace-operator/src/kubeobjects/kubernetes/statefulset/internal/types"
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 )
